+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Serve SPA on tunnel requests so PasscodeGate renders instead of raw JSON
 ---
 
 ## [0.30.0] - 2026-03-31